cluster: handle unready member
Problem statement: We could fail in MemberAdd() due to context deadline
exceeded and thus not creating related pod/service. But we could have
actually added the member. Then we would end up infinitely looping on
updateMembers() with not ready member error.

The fix is to have some intelligence when encountering unready member.
We would check the running pods and see if the member is in running
pods:
- If yes, we will take this as really not ready and back off
- otherwise, we will remove this member from membership