cluster: get name from PeerURL if member unready
Problem statement: We could fail in MemberAdd() due to context deadline
exceeded and thus not creating related pod/service. But we could have
added the member. Then we would end up infinitely looping on updateMembers()
on "not ready member" error.

One idea is that we can get name from member's PeerURL.

For self hosted case, it's more complicated because it's host IP.
We currently loop wait until the member has been added successfully.
In the future we could use pod IP to match.