cluster: remove panic
After second thought, I believe we should simply block the
operator if one of the cluster cannot make progress.

It is either caused by a bug in client-go (unlimited request waiting)
or in our code (forget to set timeout?). If that happens, the best solution
(or simplest solution) would be restarted the operator. That should be
handled at the controller layer.

So in short, we should just let it blocking when it blocks.