k8sutil: make pod annotations configurable
Annotations are a way to attach metadata to Kubernetes objects. This
PR extends the operator configuration with an option to attach
annotations to pods that the operator creates for the cluster. A
use case for this is prometheus using annotations to identify pods from
which metrics can be scraped, including pod-specific scraping
configuration. The changes in this PR would enable users to configure
etcd pods created by the operator to be annotated for scraping by
prometheus.