Add a time-out for waiting on DNS
Adding a timeout for waiting on the DNS entry to exists. This prevents the
init container to be in an inifinite loop in case of a failure of registering
the DNS on startup. This can be configured in the DNSTimeoutInSecond in
PodPolicy. The default is 0 which doesn't timout the DNS check.