Support arbitrary etcd image names
We allow a simple template to be defined in the PodPolicy, and we
perform substitution of the repository and version into that template.
The default template keeps the existing behaviour, so no user changes
should be needed.

This allows for registries that may not follow the same format as
quay.io - for example when mirroring to a private repo for no-internet
installs, the limited depth allowed by docker often means that users
will embed some of the repo name into the image name.