backup operator: fixed incorrect backup file naming convention.
When limiting the number of backups, if the revision would move to needing another digit, the logic would never select a new entry as _ is > to 9 in ascii terms.

Note that this is backwards incompatible as now files receive a correct padding and old revisions will still be incorrectly handled.