# This is a combination of 3 commits. # This is the 1st commit message:
Allow specifying Cluster Domain for clientURLs

This allows setting the ClusterDomain config in the PodPolicy. The
cluster domain is used as a suffix in the Client URLs for the etcd
members.

The ability to set a custom cluster domain is desirable when running in
clusters with a custom DNS configuration.

Signed-off-by: Mikkel Oscar Lyderik Larsen <mikkel.larsen@zalando.de>